# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SaaS POS Platform API
  version: 0.1.0
  description: |
    REST API for the SaaS POS platform. Serves two frontends:
    - **Tenant app** (`apps/web`) via `/v1/*` routes with `TenantAuthGuard`
    - **Super-admin app** (`apps/admin`) via `/admin/v1/*` routes with `AdminAuthGuard`

    Routes under `/v1/*` are tenant-scoped (RLS enforced via PostgreSQL session variable `app.current_tenant_id`).
    Routes under `/admin/v1/*` bypass RLS (set `app.is_super_admin = true`) and access cross-tenant data.

    See `architecture.md`, `admin-app.md`, and `CLAUDE.md` for the full design.

    **This file is a stub.** Phase 1 endpoints will be filled in as they are built.
    Phase 4 will publish a public API for tenants — that will be a separate, versioned doc.
  contact:
    name: Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.yourpos.com
    description: Production
  - url: https://api-staging.yourpos.com
    description: Staging
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/auth/login:
    post:
      tags: [auth-tenant]
      summary: Login (tenant user)
      operationId: loginTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful (or MFA challenge)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthResponse'
                  - $ref: '#/components/schemas/MfaChallenge'

  # ----------------------------------------------------------------------
  # Admin auth (separate realm)
  # ----------------------------------------------------------------------
  /admin/v1/auth/login:
    post:
      tags: [auth-admin]
      summary: Login (super-admin)
      operationId: loginAdmin
      description: Always returns MFA challenge first. MFA is mandatory for all super-admins.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: MFA challenge issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaChallenge'

  # ----------------------------------------------------------------------
  # Sales (tenant)
  # ----------------------------------------------------------------------
...
  /v1/products:
    get:
      tags: [inventory]
      summary: List products
      operationId: listProducts
      security:
        - TenantBearer: []
      responses:
        '200':
          description: List of products
    post:
      tags: [inventory]
      summary: Create product
      operationId: createProduct
      security:
        - TenantBearer: []
      responses:
        '201':
          description: Product created
...
components:
  securitySchemes:
    TenantBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Tenant user JWT. Validated by TenantAuthGuard.
        Sets PostgreSQL session var `app.current_tenant_id`.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Super-admin JWT. Validated by AdminAuthGuard.
        Sets PostgreSQL session var `app.is_super_admin = true`.
        Different signing secret from tenant JWTs.
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string }
            role: { type: string }
            locale: { type: string }
    MfaChallenge:
      type: object
      properties:
        challengeToken: { type: string }
        method: { type: string, enum: [totp] }

    # Money
...